Convert click outside listener to custom hook
